+++ package.json
@@ -7,7 +7,7 @@
     "@chakra-ui/react": "^1.7.2",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
-    "@undp_sdg_ai_lab/undp-radar": "^1.0.4-rc1",
+    "@undp_sdg_ai_lab/undp-radar": "^1.0.6",
     "framer-motion": "^4",
     "node-sass": "^7.0.1",
     "react": "^17.0.2",
