+++ src/pages/views/OutletContext.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const OutletContext = React.createContext<{
+  quadrant: string | undefined;
+}>({
+  quadrant: undefined
+});
